+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add pause and resume AppHangTracking API (#4077). You can now pause and resume app hang tracking with `SentrySDK.pauseAppHangTracking()` and `SentrySDK.resumeAppHangTracking()`.
 
+### Fixes
+
+- `storeEnvelope` ends session for unhandled errors (#4073)
+
 ## 8.29.1
 
 ### Fixes